pass chooser using hammerspoon

+++ hammerspoon.lua
@@ -1,17 +1,42 @@
 hs.loadSpoon('ReloadConfiguration')
 spoon.ReloadConfiguration:start()
 
-local darkmode_watcher = hs.distributednotifications.new(function(name, object, userInfo)
+local darkmode_watcher = hs.distributednotifications.new(function()
   local _, darkmode, _ = hs.osascript.applescript(
     'tell application "System Events" to tell appearance preferences to return dark mode')
-  local file = io.open(os.getenv('HOME') .. '/.config/helix/themes/theme.toml', "w")
-  if darkmode then
-    file:write('inherits = "catppuccin_mocha"')
-  else
-    file:write('inherits = "catppuccin_latte"')
+  local file = io.open(os.getenv('HOME') .. '/.config/helix/themes/theme.toml', 'w')
+  if (file) then
+    if darkmode then
+      file:write('inherits = "catppuccin_mocha"')
+    else
+      file:write('inherits = "catppuccin_latte"')
+    end
+    file:close()
+    hs.execute('pkill -USR1 hx')
   end
-  file:close()
-  hs.execute("pkill -USR1 hx")
 end, 'AppleInterfaceThemeChangedNotification')
 
 darkmode_watcher:start()
+
+local pass_chooser = hs.chooser.new(function(choice)
+  if (choice) then
+    hs.execute('gopass show --clip ' .. choice.text, true)
+  end
+end)
+
+pass_chooser:choices(function()
+  local out, status = hs.execute('gopass ls --flat', true)
+  if (status) then
+    local t = {}
+    for str in string.gmatch(out, '([^%s]+)') do
+      table.insert(t, { ['text'] = str })
+    end
+    return t
+  else
+    return 'No passwords...'
+  end
+end)
+
+hs.hotkey.bind({ 'cmd', 'ctrl' }, 'P', function()
+  pass_chooser:show()
+end)
